move files into correct folders (#70)
