+++ src/App.tsx
@@ -4,16 +4,16 @@ import { ReactLenis } from '@studio-freight/react-lenis';
 import { AnimatePresence } from 'framer-motion';
 import { Route, Routes, useLocation } from 'react-router-dom';
 import { StyleSheetManager, ThemeProvider } from 'styled-components';
-import '../src/components/ModelViewer.css';
+import './theme/ModelViewer.css';
 import Footer from './components/Footer';
 import Header from './components/Header';
-import { darkMode, lightMode } from './components/Themes';
-import { GlobalStyles } from './components/globalStyles';
+import { darkMode, lightMode } from './theme/Themes';
+import { GlobalStyles } from './theme/globalStyles';
 import { CartProvider } from './contexts/CartContext';
 import { OrderProvider } from './contexts/OrderContext';
 import { ProductProvider } from './contexts/ProductContext';
 import { useLocalStorageState } from './hooks/useLocalStorage';
-import './index.css';
+import './theme/index.css';
 import AboutUsPage from './pages/AboutUsPage';
 import BespokeDetailsPage from './pages/BespokeDetailsPage';
 import CartPage from './pages/CartPage';
