+++ src/pages/AboutUsPage.tsx
@@ -1,5 +1,5 @@
-import MarginTopContainer from '../components/MarginTopContainer';
-import PageContentWrapper from '../components/PageContentWrapper';
+import MarginTopContainer from '../components/reusable components/MarginTopContainer';
+import PageContentWrapper from '../components/reusable components/PageContentWrapper';
 
 function AboutUsPage() {
   return (
